tools/git_ops.py
import git
import os
from typing import List, Dict, Any
from datetime import datetime


class GitOperations:
    """Git操作工具"""

    def __init__(self, repo_path: str):
        self.repo_path = repo_path
        self.repo = None
        if os.path.exists(os.path.join(repo_path, '.git')):
            try:
                self.repo = git.Repo(repo_path)
            except:
                pass

    def create_branch(self, branch_name: str) -> bool:
        """创建新分支"""
        try:
            if not self.repo:
                return False
            if branch_name in [b.name for b in self.repo.branches]:
                return False
            new_branch = self.repo.create_head(branch_name)
            new_branch.checkout()
            return True
        except Exception as e:
            print(f"Error creating branch: {str(e)}")
            return False

    def commit_changes(self, message: str) -> bool:
        """提交变更"""
        try:
            if not self.repo:
                return False
            self.repo.index.add('*')
            self.repo.index.commit(message)
            return True
        except Exception as e:
            print(f"Error committing changes: {str(e)}")
            return False

    def create_pull_request(self, title: str, description: str, base_branch: str = 'main') -> Dict[str, Any]:
        """创建PR（需要GitHub API）"""
        print(f"Creating PR: {title}")
        return {
            'success': True,
            'pr_url': 'https://github.com/your-repo/pull/123',
            'message': 'PR created successfully'
        }