requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
gitpython>=3.1.40
pylint>=3.0.0
radon>=6.0.1
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
pyyaml>=6.0.1
colorlog>=6.7.0